All test commands (poetry, pytest, mypy) needed in allowlist_externals
Ref. https://stackoverflow.com/a/47716994/2664134

This resolves the tox warnings about `pytest` and `mypy`

```
WARNING: test command found but not installed in testenv
  cmd: /Users/peter_v/Library/Caches/pypoetry/virtualenvs/kopylot-1Dw6WV4w-py3.11/bin/pytest
  env: /Users/peter_v/data/ai/git/github/petervandenabeele/kopylot/.tox/py311
Maybe you forgot to specify a dependency? See also the allowlist_externals envconfig setting.

DEPRECATION WARNING: this will be an error in tox 4 and above!
```

and

```
WARNING: test command found but not installed in testenv
  cmd: /Users/peter_v/Library/Caches/pypoetry/virtualenvs/kopylot-1Dw6WV4w-py3.11/bin/mypy
  env: /Users/peter_v/data/ai/git/github/petervandenabeele/kopylot/.tox/py311
Maybe you forgot to specify a dependency? See also the allowlist_externals envconfig setting.

DEPRECATION WARNING: this will be an error in tox 4 and above!
```